+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Patch
+
+- Made build hash label clickable in BuildProgress component to navigate to build detail page
+
 ## [0.1.0] - 2025-12-10
 
 Initial release